refactored the files folder into the util folder
