+++ src/main/java/autocadDrawingChecker/util/FileChooserUtil.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.files;
+package autocadDrawingChecker.util;
 
 import autocadDrawingChecker.logging.Logger;
 import java.io.BufferedWriter;
